SLTS-13 new DataSourceListController
